Move dataclasses to independent file.
